backend/requirements.txt
fastapi==0.111.0
matplotlib==3.8.0
numpy==1.26.4
pandas==2.1.4
scikit_learn==1.2.2
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
xgboost==2.0.3
mlflow==2.12.1
psutil==5.9.0
python-dotenv

